[ISSUE-1462] Remove dependency on jmap-extensions
