+++ tmail-backend/jmap/extensions/src/main/scala/com/linagora/tmail/james/jmap/model/CalendarEventParseResponse.scala
@@ -1,0 +1,9 @@
+package com.linagora.tmail.james.jmap.model
+
+import org.apache.james.jmap.core.AccountId
+import org.apache.james.jmap.mail.BlobId
+
+case class CalendarEventParseResponse(accountId: AccountId,
+                                      parsed: Option[Map[BlobId, CalendarEventParsedList]],
+                                      notFound: Option[CalendarEventNotFound],
+                                      notParsable: Option[CalendarEventNotParsable])
