+++ tmail-backend/tmail-third-party/openpaas/pom.xml
@@ -38,10 +38,6 @@
             <artifactId>apache-james-backends-rabbitmq</artifactId>
         </dependency>
         <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>jmap-extensions</artifactId>
-        </dependency>
-        <dependency>
             <groupId>${james.groupId}</groupId>
             <artifactId>apache-james-backends-rabbitmq</artifactId>
             <scope>test</scope>
